docs : move DTO
