--- src/main/java/com/sprint/mission/discodeit/dto/UserUpdateResponse.java
@@ -1,4 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-public class UserUpdateResponse {
-}
