+++ src/main/java/com/sprint/mission/discodeit/dto/user/UserUpdateResponse.java
@@ -1,0 +1,4 @@
+package com.sprint.mission.discodeit.dto.user;
+
+public class UserUpdateResponse {
+}
